+++ src/Stack/Config.hs
@@ -185,6 +185,10 @@ configFromConfigMonoid configStackRoot configUserConfigPath mproject configMonoi
          configGhcOptions = configMonoidGhcOptions
          configSetupInfoLocations = configMonoidSetupInfoLocations
          configPvpBounds = fromMaybe PvpBoundsNone configMonoidPvpBounds
+         configModifyCodePage = fromMaybe True configMonoidModifyCodePage
+         configExplicitSetupDeps = configMonoidExplicitSetupDeps
+         configRebuildGhcOptions = fromMaybe False configMonoidRebuildGhcOptions
+         configApplyGhcOptions = fromMaybe AGOLocals configMonoidApplyGhcOptions
 
      return Config {..}
 
@@ -322,6 +326,7 @@ loadBuildConfig mproject config mresolver = do
                            , projectExtraDeps = mempty
                            , projectFlags = mempty
                            , projectResolver = r
+                           , projectExtraPackageDBs = []
                            }
                    liftIO $ do
                        S.writeFile dest' $ S.concat
@@ -355,12 +360,15 @@ loadBuildConfig mproject config mresolver = do
                 return $ mbpCompilerVersion mbp
             ResolverCompiler wantedCompiler -> return wantedCompiler
 
+    extraPackageDBs <- mapM parseRelAsAbsDir (projectExtraPackageDBs project)
+
     return BuildConfig
         { bcConfig = config
         , bcResolver = projectResolver project
         , bcWantedCompiler = wantedCompiler
         , bcPackageEntries = projectPackages project
         , bcExtraDeps = projectExtraDeps project
+        , bcExtraPackageDBs = extraPackageDBs
         , bcStackYaml = stackYamlFP
         , bcFlags = projectFlags project
         , bcImplicitGlobal = isNothing mproject